==> 1_section.ipynb <==
# Чтение данных, фильтрация дубликатов и тп

План:
- Объединим обе датасета

import pandas as pd
import numpy as np

## Объединение датасетов

business_df = pd.read_csv('business.csv')
economy_df = pd.read_csv('economy.csv')

Использование np.hstack

[
    [0, 1],
    [1, 2],
]

[
    [2, 5],
    [8, 10],
]

horizontal stack

[
    [0, 1, 2, 6],
    [1, 2, 8, 10],
]

vertical stack

[
    [0, 1],
    [1, 2],
    [2, 5],
    [8, 10],
]

general_values = np.vstack((business_df, economy_df))

# array -> dataframe перевести

general_df = pd.DataFrame(general_values, columns=business_df.columns)

Использование pd.append

general_df = business_df.append(economy_df).reset_index(drop=True)

Добавим новую колонку, которая будет обозначать какой был тип (эконом или бизнец) у билета

is_econom = [0] * business_df.shape[0] + [1] * economy_df.shape[0]
general_df['is_econom'] = is_econom

Приведем колонку stop к одному виду. Так как очень много '1-stop' без указания города транзита, то не будем смотреть на город транзита и просто сохраним три класса: '1-stop', 'non-stop', '2+-stop'

clear_stop = []

for stop in general_df.stop:
    if stop[0] == '1':
        clear_stop.append('1-stop')
    elif stop[0] == '2':
        clear_stop.append('2+-stop')
    else:
        clear_stop.append('0-stop')

# как стало

clear_stop = general_df.stop.str.strip().str[:8].str.strip()

general_df['stop'] = clear_stop

Уберем дубликаты

prev_count = general_df.shape[0]
general_df = general_df.drop_duplicates()
print('Всего дубликатов было: ', prev_count - general_df.shape[0])

Приведем price к числову виду

prices = []

for price in general_df.price:
    price = price.replace(',', '')
    prices.append(int(price))

# как стало

prices = general_df.price.str.replace(',', '').astype(int).values

general_df['price'] = prices

taken_times = []

for time in general_df.time_taken:
    a, b = time.split()

    a = a[:-1]
    b = b[:-1]

    a = float(a) * 60
    
    if b == '':
        b = 0
    else:
        b = float(b)

    taken_times.append(a + b)

# как стало

def process_str_time(str_time):
    a, b = list(map(lambda x: x[:-1], str_time.split()))
    
    if b:
        return int(a) * 60 + int(b)
    
    return float(a) * 60

taken_times = general_df.time_taken.apply(process_str_time).values

general_df['taken_time_minues'] = taken_times

Избавимся от ch_code, num_code, потому что .... И от arr_time, потому что мы знаем время в полете и когда самолет вылетел и мы сможем его восстановить. time_taken тоже, потому что мы его перевели в число минут

result_df = general_df.drop(['ch_code', 'num_code', 'arr_time', 'time_taken'], axis=1)

result_df.to_csv('features.csv', index=False)

result_df.head()
