==> 2_section.ipynb <==
# Feature Engineering

- Координаты городов. Чтобы в основном найти расстояния.
- Перелеты внутри государства (расстояние между городами). Скорее всего цена зависит от расхода топлива -> от расстояния
- Кластеризация.
- Дата: выходной.
- Скорость: расстояние / время полета.
- Часто встречаемые авиалинии.
- Погода.
- Время полета (день, утро, вечер).
- В какую минуту дня самолет вылетел.

import pandas as pd

df = pd.read_csv('features.csv', parse_dates=['date'], sep=',')

### Координаты городов

city_name -> [lat, lon]

str -> [float, float]

москва -> [63, 128]

from mapbox import Geocoder

access_token = 'pk.eyJ1IjoiZWdvcm92bSIsImEiOiJjanhic2F1ZDUwMTdvNDJ0bHMxOGg5NWthIn0.QL2wIcgAJIBNIz2gbpuLVQ'

geocoder = Geocoder(access_token=access_token)

all_cities = df['from'].append(df['to']).unique()

city_coordinate = {}

for city_name in all_cities:
    response = geocoder.forward(city_name)
    place_name = response.json()['features'][0]['place_name']
    coordinates = response.json()['features'][0]['geometry']['coordinates']
    
    city_coordinate[city_name] = coordinates

city_coordinate

from_coordinates_lat = []
from_coordinates_lon = []

for city_name in df['from']:
    from_coordinates_lat.append(city_coordinate[city_name][1]) # lat
    from_coordinates_lon.append(city_coordinate[city_name][0]) # lon
    
    
to_coordinates_lat = []
to_coordinates_lon = []

for city_name in df['to']:
    to_coordinates_lat.append(city_coordinate[city_name][1])
    to_coordinates_lon.append(city_coordinate[city_name][0])

df['from_coordinates_lat'] = from_coordinates_lat
df['from_coordinates_lon'] = from_coordinates_lon

df['to_coordinates_lat'] = to_coordinates_lat
df['to_coordinates_lon'] = to_coordinates_lon

import haversine

# haversine.haversine((1, 1), (2, 4))

distances = []

for i in range(len(from_coordinates_lat)):
    from_lat = from_coordinates_lat[i]
    from_lon = from_coordinates_lon[i]
    
    to_lat = to_coordinates_lat[i]
    to_lon = to_coordinates_lon[i]
    
    distance = haversine.haversine((from_lat, from_lon), (to_lat, to_lon))
    
    distances.append(distance)

df['distance'] = distances

df['speed'] = df['distance'] / (df['taken_time_minues'] / 60)

df['week_day'] = df['date'].dt.weekday

df.head()

df['dep_minutes'] = df.dep_time.apply(lambda x: int(x.split(':')[0]) * 60 + int(x.split(':')[1]))

df['stop_count'] = df.stop.map({'non-stop': 0, '1-stop': 1, '2+-stop': 2})

new_airlines = []

for airline in df.airline:
    if airline != 'Vistara' and airline != 'Air India' and airline != 'Indigo':
        new_airlines.append('Other')
    else:
        new_airlines.append(airline)

df['new_airlines'] = new_airlines

df.to_csv('new_features.csv', index=False)
